refactor: modularize main.py into cli/ package
Split the monolithic main.py (1970 lines) into a modular cli/ package:

- cli/__init__.py - Package exports and main() entry point
- cli/parser.py - Argument parsing and config resolution
- cli/system.py - LeapTradingSystem orchestrator class
- cli/commands/ - Command handlers (train, backtest, walkforward, evaluate, autotrade)

Maintains full backward compatibility:
- `from main import LeapTradingSystem` still works
- Test patching patterns like `@patch('main.TransformerPredictor')` still work
- All CLI commands function identically

Documentation:
- Added ADR-0008 documenting the refactoring decision
- Updated ARCHITECTURE.md with new directory structure
- Added cli/CLAUDE.md for module-specific patterns
